# frontend/public/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cursor AI Full-Stack Application API",
    "description": "REST API for the Cursor AI application including Customer Support Ticket System and Blog Platform.\n\n## Features\n- **Authentication**: JWT-based authentication with refresh tokens\n- **Tickets**: Complete ticket management system with SLA tracking\n- **Blog**: Full-featured blog with posts, comments, and categories\n- **Users**: User management with role-based access control\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your_token>\n```",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/dnistrova/cursor-ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "per_page", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/UserList"}
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "User data"},
          "404": {"description": "User not found"}
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/UserUpdate"}
            }
          }
        },
        "responses": {
          "200": {"description": "User updated"},
          "403": {"description": "Forbidden"}
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}
        ],
        "responses": {
          "204": {"description": "User deleted"},
          "403": {"description": "Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "email": {"type": "string"},
          "username": {"type": "string"},
          "first_name": {"type": "string"},
          "last_name": {"type": "string"},
          "role": {"type": "string", "enum": ["customer", "agent", "admin"]},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "first_name": {"type": "string"},
          "last_name": {"type": "string"},
          "avatar_url": {"type": "string"}
        }
      },
      "UserList": {
        "type": "object",
        "properties": {
          "users": {"type": "array", "items": {"$ref": "#/components/schemas/User"}},
          "total": {"type": "integer"},
          "page": {"type": "integer"},
          "per_page": {"type": "integer"}
        }
      },
...
    }
  }
}